исправлено пересечение отрезков(совсем другой подход)

+++ other_func.cpp
@@ -12,7 +12,7 @@ double det(double a, double b, double c, double d) {
     return a * d - b * c;
 }
 
-double orientation(Point first_point, Point second_point, Point third_point) {
+int orientation(Point first_point, Point second_point, Point third_point) {
 
     double val = (second_point.y() - first_point.y()) * (third_point.x() - second_point.x()) -
                  (second_point.x() - first_point.x()) * (third_point.y() - second_point.y());
